Check before removing framework folder

+++ FLIMfitFrontEnd/compile.m
@@ -112,7 +112,9 @@ function compile(exit_on_error)
                 case 'MAC'
 
                     % Manually fix Qt references
-                    rmdir('Libraries/QtCore.framework','s')
+                    if isdir('Libraries/QtCore.framework')
+                        rmdir('Libraries/QtCore.framework','s')
+                    end
                     copyfile('/usr/local/opt/qt/lib/QtCore.framework/','Libraries/QtCore.framework/');
                     system('install_name_tool -change /usr/local/opt/qt/lib/QtCore.framework/Versions/5/QtCore @loader_path/QtCore.framework/Versions/5/QtCore Libraries/FLIMfitMex.mexmaci64')
